+++ .gitignore
@@ -20,3 +20,7 @@ src/NuGetBuild/
 src/NuGetBuildWp7/
 *.nupkg
 README.html
+/src/.vs/SettingsProviderNet/v16/Server/sqlite3/storage.ide-wal
+/src/.vs/SettingsProviderNet/v16/Server/sqlite3/storage.ide-shm
+/src/.vs/SettingsProviderNet/v16/Server/sqlite3/storage.ide
+/src/.vs/SettingsProviderNet/v16/Server/sqlite3/db.lock
