+++ src/SettingsProviderNet/SettingsProviderNet.csproj
@@ -44,13 +44,13 @@
     <Compile Include="ISettingDescriptor.cs" />
     <Compile Include="ISettingsProvider.cs" />
     <Compile Include="ISettingsStorage.cs" />
-    <Compile Include="IsolatedStorageSettingsStore.cs" />
-    <Compile Include="JsonSettingsStoreBase.cs" />
+    <Compile Include="Storages\IsolatedStorageSettingsStore.cs" />
+    <Compile Include="Storages\JsonSettingsStoreBase.cs" />
     <Compile Include="KeyAttribute.cs" />
-    <Compile Include="PortableStorage.cs" />
+    <Compile Include="Storages\PortableStorage.cs" />
     <Compile Include="ProtectedDataUtils.cs" />
     <Compile Include="ProtectedStringAttribute.cs" />
-    <Compile Include="RoamingAppDataStorage.cs" />
+    <Compile Include="Storages\RoamingAppDataStorage.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SettingDescriptor.cs" />
     <Compile Include="SettingsProvider.cs" />
