+++ src/SettingsProviderNet/Storages/JsonSettingsStoreBase.cs
