+++ src/SettingsProviderNet/Storages/PortableStorage.cs
