+++ src/SettingsProviderNet/Storages/RoamingAppDataStorage.cs
